requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
langchain>=0.2.0
google-generativeai>=0.7.0
chromadb>=0.5.0
pymupdf>=1.24.0
python-dotenv>=1.0.1
langchain-google-genai>=1.0.7
langchain-groq>=0.1.3
pytesseract>=0.3.10
Pillow>=10.3.0
streamlit>=1.36.0
pdfplumber>=0.11.4
opencv-python-headless>=4.10.0
